requirements.txt
# Python依赖包
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.20.0
PyYAML>=5.4
Pillow>=9.0.0
scikit-image>=0.19.0
scipy>=1.7.0
tensorboard>=2.11.0

# 可选依赖
# wandb>=0.12.0  # 用于Weights & Biases日志
# opencv-python>=4.5.0  # 用于图像处理